Fixed missing default bug
Added fix and tests that make sure when no default profile is in the file to begin with, a default profile is added and defaulted to the top profile found.
Added dummy files so that the `Should find the credentials file in a directory` test also covers avoiding other files.
#2